requirements.txt
greenlet==2.0.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.17
tqdm==4.65.0
typing_extensions==4.7.1
